refactor(LG-39): API 공통 응답 클래스를 생성한다
