+++ src/main/java/com/linkgem/presentation/common/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package com.linkgem.presentation.exception;
+package com.linkgem.presentation.common.exception;
 
 import lombok.Getter;
