+++ src/main/java/com/linkgem/presentation/common/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.linkgem.presentation.exception;
+package com.linkgem.presentation.common.exception;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
